move v2 test into collation suite
